Bug 2028887 - Use 64-bit arithmetic. r=gfx-reviewers,lsalzman
(As suggested by clauditor.)

Differential Revision: https://phabricator.services.mozilla.com/D291997

UltraBlame original commit: 40cff7cd0b75e40fdfcba52583d6f0c4e7fe15b9

+++ gfx/thebes/COLRFonts.cpp
@@ -2,7 +2,6 @@
 
 
 
-
 #include "COLRFonts.h"
 #include "gfxFontUtils.h"
 #include "gfxUtils.h"
@@ -2336,7 +2335,7 @@ bool ItemVariationData::Validate(const COLRv1Header* aHeader,
   uint32_t deltaSetSize =
       (uint16_t(regionIndexCount) + uint16_t(wordDeltaCount)) << longWords;
   if (reinterpret_cast<const char*>(deltaSets()) +
-          uint16_t(itemCount) * deltaSetSize >
+          uint64_t(uint16_t(itemCount)) * deltaSetSize >
       reinterpret_cast<const char*>(aHeader) + aLength) {
     return false;
   }
